@@ -1,14 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion '28.0.3'
+    compileSdkVersion 32
     defaultConfig {
         applicationId "innatemobile.storymakerevents"
         minSdkVersion 15
-        targetSdkVersion 28
-        versionCode 19
-        versionName '1.20'
+        targetSdkVersion 32
+        versionCode 20
+        versionName '1.14'
     }
     buildTypes {
         release {
@@ -23,10 +22,10 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     testImplementation 'junit:junit:4.13'
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support:design:28.0.0'
-    implementation 'com.android.support:cardview-v7:28.0.0'
-    implementation 'com.android.support:support-v4:28.0.0'
+    implementation 'androidx.appcompat:appcompat:1.4.1'
+    implementation 'com.google.android.material:material:1.6.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation 'com.cloudinary:cloudinary-android:1.3.0'
     implementation 'com.mcxiaoke.volley:library:1.0.19'
 }
