@@ -2,15 +2,16 @@ package innatemobile.storymakerevents.Activities;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
 import android.view.MenuItem;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
 import java.util.List;
 import innatemobile.storymakerevents.Adapters.AddScheduleAdapter;
 import innatemobile.storymakerevents.Models.ScheduleJoined;
-import innatemobile.storymakerevents.Models.Schedules;
 import innatemobile.storymakerevents.R;
 import innatemobile.storymakerevents.Utils.AppController;
 import innatemobile.storymakerevents.Utils.DatabaseHandler;
@@ -101,9 +102,5 @@ public class AddScheduleActivity extends AppCompatActivity implements AddSchedul
         adapter.notifyItemChanged(position);
     }
 
-    @Override
-    public void removedPresentation(int position) {
-        adapter.notifyItemChanged(position);
-    }
     /***********************INTERFACE METHODS****************************/
 }
