@@ -1,11 +1,13 @@
 package innatemobile.storymakerevents.Adapters;
 
 import android.app.Activity;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
+
 import java.util.List;
 import innatemobile.storymakerevents.Models.Breakouts;
 import innatemobile.storymakerevents.R;
